server/requirements.txt
openenv-core[core]>=0.2.3
click>=8.1.8,<9.0
fastapi>=0.115.2,<1.0
uvicorn[standard]>=0.35.0,<1.0
pydantic>=2.11.7,<3.0
httpx>=0.28.1,<1.0
